Fixed the number of leds being an uint8_t limiting max number of leds.  Closes: #4

+++ src/ws2812.h
@@ -5,7 +5,7 @@
  ******************************************************************************
  * @attention
  *
- * Copyright (c) 2022 - 2025 Lars Boegild Thomsen <lbthomsen@gmail.com>
+ * Copyright (c) 2022 - 2026 Lars Boegild Thomsen <lbthomsen@gmail.com>
  * All rights reserved.
  *
  * This software is licensed under terms that can be found in the LICENSE file
@@ -52,7 +52,7 @@ typedef struct {
     uint16_t leds;                          // Number of LEDs on the string
     uint8_t *led;                           // Dynamically allocated array of LED RGB values
     ws2812_stateTypeDef led_state;          // LED Transfer state machine
-    uint8_t led_cnt;                        // Counts through the leds starting from zero up to "leds"
+    uint16_t led_cnt;                        // Counts through the leds starting from zero up to "leds"
     uint8_t res_cnt;                        // Counts reset cycles when in reset state
     uint8_t is_dirty;                       // Indicates to the call back that the led color values have been updated
     uint8_t zero_halves;                    // Counts halves send during reset
